chore(EMS-2727): no pdf - split up keystone api tests
