+++ src/api/keystone-create-application.test.ts
@@ -1,16 +1,13 @@
 import { addMonths } from 'date-fns';
 import { APPLICATION } from './constants';
-import updateApplication from './helpers/update-application';
-import accounts from './test-helpers/accounts';
 import getKeystoneContext from './test-helpers/get-keystone-context';
 import applications from './test-helpers/applications';
 import buyers from './test-helpers/buyers';
 import buyerTradingHistoryHelper from './test-helpers/buyer-trading-history';
 import policies from './test-helpers/policies';
-import { mockAccount } from './test-mocks';
-import { Application, ApplicationBuyer, ApplicationBusiness, Account, Context } from './types';
+import { Application, ApplicationBuyer, ApplicationBusiness, Context } from './types';
 
-describe('Create an Application', () => {
+describe('Keystone - Create an Application', () => {
   let context: Context;
   let application: Application;
 
@@ -247,224 +244,5 @@ describe('Create an Application', () => {
     });
 
     expect(declaration.application.id).toEqual(application.id);
-  });
-});
-
-describe('Account', () => {
-  let context: Context;
-  let account: Account;
-
-  describe('create', () => {
-    beforeAll(async () => {
-      context = getKeystoneContext();
-
-      await accounts.create({ context });
-    });
-
-    describe('when an account already exists with the provided email', () => {
-      test('it should not create the account', async () => {
-        const response = (await accounts.create({ context, data: mockAccount, deleteAccounts: false })) as Account;
-
-        expect(response.id).toBeUndefined();
-        expect(response.email).toBeUndefined();
-
-        const allAccounts = await context.query.Account.findMany();
-
-        expect(allAccounts.length).toEqual(1);
-      });
-    });
-  });
-
-  describe('update', () => {
-    let updatedAccount: Account;
-
-    const accountUpdate = { firstName: 'Updated' };
-
-    beforeAll(async () => {
-      account = await accounts.create({ context });
-
-      updatedAccount = (await context.query.Account.updateOne({
-        where: { id: account.id },
-        data: accountUpdate,
-        query: 'id createdAt updatedAt firstName lastName email salt hash isVerified verificationHash verificationExpiry',
-      })) as Account;
-    });
-
-    test('it should update the provided fields', () => {
-      expect(updatedAccount.firstName).toEqual(accountUpdate.firstName);
-    });
-
-    test('it should update updatedAt', () => {
-      expect(updatedAccount.updatedAt).not.toEqual(account.createdAt);
-    });
-  });
-});
-
-describe('Application timestamp updates', () => {
-  let context: Context;
-  let application: Application;
-
-  const updateApplicationTimestampSpy = jest.fn();
-
-  beforeAll(async () => {
-    context = getKeystoneContext();
-
-    application = (await applications.create({ context, data: {} })) as Application;
-
-    // create buyer and associate with the application.
-    const buyer = await buyers.create({
-      context,
-      data: {
-        application: {
-          connect: { id: application.id },
-        },
-      },
-    });
-
-    /**
-     * Create buyer trading history,
-     * Associate with the buyer and application
-     */
-    const buyerTradingHistory = await buyerTradingHistoryHelper.create({
-      context,
-      data: {
-        buyer: {
-          connect: { id: buyer.id },
-        },
-        application: {
-          connect: { id: application.id },
-        },
-      },
-    });
-
-    // create policy and associate with the application.
-    const policy = await policies.create({
-      context,
-      data: {
-        application: {
-          connect: { id: application.id },
-        },
-      },
-    });
-
-    // create company and associate with the application.
-    const company = await context.query.Company.createOne({
-      data: {
-        application: {
-          connect: { id: application.id },
-        },
-      },
-    });
-
-    application = {
-      ...application,
-      buyer: {
-        ...buyer,
-        buyerTradingHistory,
-      },
-      company: {
-        id: company.id,
-      },
-      policy: {
-        id: policy.id,
-      },
-    };
-  });
-
-  beforeEach(() => {
-    jest.resetAllMocks();
-
-    jest.mock('./helpers/update-application');
-
-    updateApplication.timestamp = updateApplicationTimestampSpy;
-  });
-
-  const assertSpyWasCalled = () => {
-    expect(updateApplicationTimestampSpy).toHaveBeenCalledTimes(1);
-    expect(updateApplicationTimestampSpy).toHaveBeenCalledWith(context, application.id);
-  };
-
-  describe('Policy', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Policy.updateOne({
-        where: { id: application.policy.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('Business', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Business.updateOne({
-        where: { id: application.business.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('Broker', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Broker.updateOne({
-        where: { id: application.broker.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('Company', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Company.updateOne({
-        where: { id: application.company.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('Buyer', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Buyer.updateOne({
-        where: { id: application.buyer.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('BuyerTradingHistory', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.BuyerTradingHistory.updateOne({
-        where: { id: application.buyer.buyerTradingHistory.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
-  });
-
-  describe('Declaration', () => {
-    test('it should call updateApplication.timestamp', async () => {
-      await context.query.Declaration.updateOne({
-        where: { id: application.declaration.id },
-        data: {},
-        query: 'id',
-      });
-
-      assertSpyWasCalled();
-    });
   });
 });
